reduce disk space by having fewer integration test files which broke CI
